% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIFpkg.R
\name{VIFsM}
\alias{VIFsM}
\alias{Plot}
\alias{summarY}
\title{-Variance Inflation Factor (VIF)-}
\usage{
VIFsM(Data, multi = FALSE, maxim = 4)

Plot(x, maxx = 4, multi = FALSE, ...)

summarY(object, multi = FALSE, ...)
}
\arguments{
\item{Data}{a dataframe or matrix of predictors}

\item{multi}{logical stating whether to drop the variable with VIF greater than maxim, and proceed to recompute VIF until no multicollinearity exist based on maxim value defined. Default is FALSE, which just computes the VIF of the current data}

\item{maxim}{numeric value. The maximum acceptable VIF value. Predictor variables are sequentially kicked out of the model if they have VIF greater than the maxim when multi = TRUE. Default is value 4.}

\item{x}{object of class \code{VIFsM}}

\item{maxx}{numeric value. The maximum acceptable VIF value. Predictor variables are sequentially kicked out of the model if they have VIF greater than the maxim when multi = TRUE. Default is value 4.}

\item{...}{Further arguments passed to ggplot2}

\item{object}{of class \code{VIFsM}}
}
\value{
A list.

a plot of the VIF values

a summary of the VIF values
}
\description{
An R package that computes the Variance Inflation Factor (VIF) of predictor variables
}
\examples{
# Using the Boston data from MASS package
df <- data.frame(MASS::Boston[,-1])
# When Interested in the VIF of the current data (multi = FALSE, as in default)
xs <- VIFsM(df, multi=FALSE,maxim = 4)
Plot(x = xs,maxx = 4)
summarY(object=xs)
# When Interested in checking the VIF of the data and dropping the variable with VIF > maxim
xs <- VIFsM(df, multi=TRUE,maxim = 4)
Plot(x = xs,multi = TRUE, maxx = 4)
summarY(object=xs, multi = TRUE)
}
